refactor(findings,drafting,baseline): TS-159 store Finding.amount_exposure in minor units (paise)
- Change amount_exposure from Numeric(16,2) float to BigInteger int minor units.

- Update app.core.contracts.findings amount_exposure type to int.

- Update drafting validators to parse and compare rupee amounts in paise.

- Remove float conversions in baseline and drafting services.

- Update tests and add Alembic migration f5c03761fb0c for PostgreSQL/SQLite.

- Update specs/modules/findings.md, tasks/backlog.md, CHANGELOG.md.

Task: TS-159

+++ backend/app/core/contracts/findings.py
@@ -61,7 +61,7 @@ class Finding(BaseModel):
     suggested_action: str | None = None
     pattern_id: str | None = None
     pattern_version: str | None = None
-    amount_exposure: float | None = None
+    amount_exposure: int | None = None
     review_status: ReviewStatus = ReviewStatus.PROPOSED
     review_reason: str | None = None
     explanation: dict | None = None
